backend/auth/requirements.txt
fastapi
uvicorn
SQLAlchemy
bcrypt
python-jose[cryptography]
passlib[bcrypt]